do not produce an error for a normal worker exits with inputs outstanding
A worker may terminate normally, yet still have inputs in its queue, if
those inputs were received after the worker was asked to shutdown (eoi).
This can happen in recursive input situations.  It's not an error
situation, so it shouldn't generate an error trace (as it did before
this commit).