per-fitting queue limits
fitting_spec.q_limit can now be set to specify a different queue-length limit
for each fitting.  The application environment variable worker_q_limit is now
a global cap on that setting.  That is, the actually queue limit used is the
lesser of #fitting_spec.q_limit and riak_pipe:worker_q_limit.