catch the exit that will happen if Node is unavailable
Node may have crashed (or otherwise become unavailable) between the time
that the preflist was generated, and the time that this call the the
vnode master is made.  If that happens, the gen_server:call will exit.
The exit is caught here to allow an attempt at another node in the
preflist.

This error came up intermittently in riak_pipe:limits_test_, which
starts and stops a slave node.  It showed up in the eunit logs as a
badmatch, because the worker process exit would generate an error trace
message, and the test expects zero error trace messages.