catch sync_send_event in worker_done in case fitting is gone
BZ1236

In timeout and other error cases that tear down the whole pipeline, it's
possible that a worker may finish its work and cause the vnode to report
back to the fitting process after that process has already exited.  This
patch catches the exit that gen_fsm:sync_send_event would otherwise force in
this case, to prevent the vnode (which is calling this method) from falling
over.