reduce uninformative process death spam
Using supervisor:terminate_child/2 instead of erlang:exit/2 means that
the supervisor doesn't complain about the untimely death of its worker
process.  Combine this with changing the fittings from linking to
monitoring of the builder process (so the builder isn't torn down
abnormally when a fitting dies), and you get the end of "fitting_died"
spam.