remove the sink_died exit reason
{sink_died, normal} is harmless, yet confusing, error log spam.  This
patch removes it by making the builder exit 'normal' when the sink
exits.

The implementation in this patch will also prevent the builder from
complaining about abnormal sink exits.  I think this is desired
behavior, because the sink and/or its supervisor should have logged such
an exit already, so there should be no need for the builder to log it
again.