sink backpressure via gen_fsm sync events
This patch allows a Pipe user to specify the "type" of the sink process.
This type is used to decide how to send it #pipe_result{} messages.

If the sink type is 'raw' or unspecified, result messages are delivered
as regular erlang messages, as they have been before this commit.

If the sink type is '{fsm_sync, Timeout}', result messages are delivered
as calls to gen_fsm:sync_send_message/3. The biggest benefit of this
sink type is that it will provide backpressure to the pipe, to prevent
the sink's mailbox from being overwhelmed.