assume handoff if vnode exits 'normal' during queue requeuest
The pipe_verify_handoff_blocking test in riak_test PR 154 demonstrated
that the monitor used by riak_pipe_vnode:queue_work_send may fire while
handoff is taking place. This situation is further described in
riak_pipe issue 63.

This commit addresses the problem by first checking whether the Reason
for the vnode's exit was 'normal'. If it was not 'normal', the
vnode_down error is returned as before. If it was 'normal', then it is
assumed that handoff happened, and the monitor is moved to the new
vnode, as indicated by riak_core_ring:next_owner/2.