broaden catch from exit:{noproc,_} to exit:_ in riak_pipe_fitting
This buglet was found by the new riak_pipe_fitting_eqc test. If the fitting was still up
when the gen_fsm message was called, but then stopped before replying, the exit reason
would be something like {normal, _} instead of the {noproc,_} already being caught.

This patch catches all exits to make sure execution ends up in the 'gone' case of
riak_pipe_vnode:new_worker, instead of in the catch clause. In the riak_pipe_fitting_eqc
test, the exit caused the test to abort, because it did not include the try/catch in its
fake vnode implementation.