first fitting pulse test to produce an error
If a fitting has no active workers when eoi arrives, it may evaluate
the worker behavior in-process. This might cause the fitting to
exit non-normal for any number of reasons, but PULSE found specifically
that the sink exiting while a synchronous message is being sent to it
(fsm-type sink, on the sync period), the gen_fsm call will exit
abnormally, but not necessarily in any of the ways already caught.
The specific exit seen here was {normal, {gen...}}.