cherry-pick commit 041cee5 into 1.3 branch
It was desired to have the sink-sync-send crash fix backported to
the 1.3 branch. This is a cherry-pick of the commit that fixed the
issue from pull request #73. The tests tests that demonstrate the
issue have not been backported, to save merge hassle.

Original commit comment:

catch sink fsm exits for all reasons

The reduce_fitting_pulse test showed that other exit reasons might
be seen here. For example, 'normal' can happen if the sink shuts
down after the sync event has been delivered.

This commit broadens the catch to prevent those other reasons from
killing the current process, and also narrows it to make sure exit
is bubbling from the sync_send_event, and nowhere else.