lower "fitting was gone" log to debug level
Issue #50 is not the only complaint about the message,
"Pipe worker startup failed:fitting was gone before startup".
It's not an important message, though. It *might* tell you that a
pipe vnode is lagging a bit, but most often it's just noise about
pipes that shut down for other reasons (MapReduce errors).

This commit lowers the log level of the message to 'debug' and
rewords it, since it really has nothing to do with the worker
failing to start.