Remove hack used to print variable types
The `type_of` function was implemented using undocumented APIs to
inspect opaque terms. Definitely not future-proof. It was also only used
to print types of variables in error messages, and in every instance it
would be more helpful to just print the contents of the variable itself.

So I updated the error messages and removed `type_of` altogether.