Some flappy test failure fixes (#113)
* fix(?) for flappy test failure

* Fix process leak and intermittent failure

Use a monitor to decide if a process is dead, rather than presence in
the `processes()` list, as intermitently the list was inconsistent. I
didn't get to the bottom of why. Using a monitor works. Also there was a
process leak with vnode processes. Not that it matters from eunit, but
running for > 10 minutes on my machine would lead to erlang running out
of processes.

* Update riak-core dep for nhs-riak

* update deps for basho-dev-2.2.5